Extract shared test fake
